+++ src/apps/curve/curve.module.ts
@@ -5,6 +5,7 @@ import { ArbitrumCurveBalanceFetcher } from './arbitrum/curve.balance-fetcher';
 import { ArbitrumCurveGaugeContractPositionFetcher } from './arbitrum/curve.gauge.contract-position-fetcher';
 import { ArbitrumCurvePoolTokenFetcher } from './arbitrum/curve.pool.token-fetcher';
 import { AvalancheCurveBalanceFetcher } from './avalanche/curve.balance-fetcher';
+import { AvalancheCurveGaugeContractPositionFetcher } from './avalanche/curve.gauge.contract-position-fetcher';
 import { AvalancheCurvePoolTokenFetcher } from './avalanche/curve.pool.token-fetcher';
 import { CurveContractFactory } from './contracts';
 import { CurveAppDefinition, CURVE_DEFINITION } from './curve.definition';
@@ -27,6 +28,8 @@ import { CurveGaugeIsActiveStrategy } from './helpers/curve.gauge.is-active-stra
 import { CurveGaugeRegistry } from './helpers/curve.gauge.registry';
 import { CurveGaugeRoiStrategy } from './helpers/curve.gauge.roi-strategy';
 import { CurvePoolDefaultTokenHelper } from './helpers/curve.pool.default.token-helper';
+import { CurvePoolOnChainCoinStrategy } from './helpers/curve.pool.on-chain.coin-strategy';
+import { CurvePoolOnChainReserveStrategy } from './helpers/curve.pool.on-chain.reserve-strategy';
 import { CurvePoolRegistry } from './helpers/curve.pool.registry';
 import { CurvePoolTokenHelper } from './helpers/curve.pool.token-helper';
 import { CurvePoolVirtualPriceStrategy } from './helpers/curve.pool.virtual.price-strategy';
@@ -53,7 +56,7 @@ import { PolygonCurvePoolTokenFetcher } from './polygon/curve.pool.token-fetcher
     // Avalanche
     AvalancheCurveBalanceFetcher,
     AvalancheCurvePoolTokenFetcher,
-    ArbitrumCurveGaugeContractPositionFetcher,
+    AvalancheCurveGaugeContractPositionFetcher,
     // Ethereum
     EthereumCurveBalanceFetcher,
     EthereumCurvePoolTokenFetcher,
@@ -76,18 +79,23 @@ import { PolygonCurvePoolTokenFetcher } from './polygon/curve.pool.token-fetcher
     PolygonCurveBalanceFetcher,
     PolygonCurvePoolTokenFetcher,
     PolygonCurveGaugeContractPositionFetcher,
-    // Token Helpers
+    // API Helpers
     CurveApiClient,
+    // Pool Token Helpers
     CurvePoolRegistry,
     CurvePoolTokenHelper,
     CurvePoolDefaultTokenHelper,
     CurvePoolVirtualPriceStrategy,
+    CurvePoolOnChainCoinStrategy,
+    CurvePoolOnChainReserveStrategy,
+    // Gauge Helpers
     CurveGaugeRegistry,
     CurveGaugeDefaultContractPositionHelper,
     CurveGaugeDefaultContractPositionBalanceHelper,
     CurveGaugeIsActiveStrategy,
     CurveGaugeRoiStrategy,
     CurveChildLiquidityGaugeRoiStrategy,
+    // Escrow Helpers
     CurveVotingEscrowContractPositionHelper,
     CurveVotingEscrowContractPositionBalanceHelper,
     CurveVestingEscrowContractPositionHelper,
@@ -95,16 +103,23 @@ import { PolygonCurvePoolTokenFetcher } from './polygon/curve.pool.token-fetcher
   ],
   exports: [
     CurveContractFactory,
+    // API Helpers
     CurveApiClient,
+    // Pool Token Helpers
     CurvePoolRegistry,
     CurvePoolTokenHelper,
     CurvePoolDefaultTokenHelper,
+    CurvePoolVirtualPriceStrategy,
+    CurvePoolOnChainCoinStrategy,
+    CurvePoolOnChainReserveStrategy,
+    // Gauge Helpers
     CurveGaugeRegistry,
     CurveGaugeDefaultContractPositionHelper,
     CurveGaugeDefaultContractPositionBalanceHelper,
     CurveGaugeIsActiveStrategy,
     CurveGaugeRoiStrategy,
     CurveChildLiquidityGaugeRoiStrategy,
+    // Escrow Helpers
     CurveVotingEscrowContractPositionHelper,
     CurveVotingEscrowContractPositionBalanceHelper,
     CurveVestingEscrowContractPositionHelper,